+++ tests/test_xterm_parser.py
@@ -33,7 +33,7 @@ def chunks(data, size):
 
 @pytest.fixture
 def parser():
-    return XTermParser(more_data=lambda: False)
+    return XTermParser()
 
 
 @pytest.mark.parametrize("chunk_size", [2, 3, 4, 5, 6])
@@ -110,13 +110,9 @@ def test_cant_match_escape_sequence_too_long(parser):
 @pytest.mark.parametrize(
     "chunk_size",
     [
-        pytest.param(
-            2, marks=pytest.mark.xfail(reason="Fails when ESC at end of chunk")
-        ),
+        2,
         3,
-        pytest.param(
-            4, marks=pytest.mark.xfail(reason="Fails when ESC at end of chunk")
-        ),
+        4,
         5,
         6,
     ],
@@ -132,14 +128,15 @@ def test_unknown_sequence_followed_by_known_sequence(parser, chunk_size):
     sequence = unknown_sequence + known_sequence
 
     events = []
-    parser.more_data = lambda: True
+
     for chunk in chunks(sequence, chunk_size):
         events.append(parser.feed(chunk))
 
     events = list(itertools.chain.from_iterable(list(event) for event in events))
+    print(repr([event.key for event in events]))
 
     assert [event.key for event in events] == [
-        "circumflex_accent",
+        "escape",
         "left_square_bracket",
         "question_mark",
         "end",
@@ -169,14 +166,17 @@ def test_key_presses_and_escape_sequence_mixed(parser):
 
 def test_single_escape(parser):
     """A single \x1b should be interpreted as a single press of the Escape key"""
-    events = parser.feed("\x1b")
+    events = list(parser.feed("\x1b"))
+    events.extend(parser.feed(""))
     assert [event.key for event in events] == ["escape"]
 
 
 def test_double_escape(parser):
-    """Windows Terminal writes double ESC when the user presses the Escape key once."""
-    events = parser.feed("\x1b\x1b")
-    assert [event.key for event in events] == ["escape"]
+    """Test double escape."""
+    events = list(parser.feed("\x1b\x1b"))
+    events.extend(parser.feed(""))
+    print(events)
+    assert [event.key for event in events] == ["escape", "escape"]
 
 
 @pytest.mark.parametrize(